Rethink how cacheing works
There is now an explicit cache, in ~/.cache, for mod version info. It is
explicitly refreshed with "ksp refresh"; other commands will populate
the cache on misses but will not update stale entries. This adds another
step to checking for updates (ksp refresh; ksp check) but hugely
speeds up repeated "ksp check" commands and reduces load on the forums.